refactor: structural stratification for workflow domain
Extract constraint and policy models from `topologies.py` and `nodes.py` into dedicated `constraints.py` and `policies.py` files to resolve the "God Model" anti-pattern. Update the public API facade in `__init__.py` to maintain the boundary.
