+++ src/coreason_manifest/workflow/nodes.py
@@ -11,6 +11,7 @@ from pydantic import Field
 
 from coreason_manifest.core.base import CoreasonBaseModel
 from coreason_manifest.oversight.intervention import InterventionPolicy
+from coreason_manifest.workflow.policies import EpistemicScanner, SelfCorrectionPolicy, System1Reflex
 
 
 class BaseNode(CoreasonBaseModel):
@@ -23,40 +24,6 @@ class BaseNode(CoreasonBaseModel):
         default_factory=list,
         description="A declarative list of proactive oversight hooks bound to this node's lifecycle.",
     )
-
-
-class System1Reflex(CoreasonBaseModel):
-    """
-    Policy for fast, intuitive system 1 thinking.
-    """
-
-    confidence_threshold: float = Field(
-        ge=0.0, le=1.0, description="The confidence threshold required to execute a reflex action."
-    )
-    allowed_read_only_tools: list[str] = Field(description="List of read-only tools allowed during a reflex action.")
-
-
-class EpistemicScanner(CoreasonBaseModel):
-    """
-    Policy for epistemic scanning and gap detection.
-    """
-
-    active: bool = Field(description="Whether the epistemic scanner is active.")
-    dissonance_threshold: float = Field(
-        ge=0.0, le=1.0, description="The threshold for cognitive dissonance before triggering an action."
-    )
-    action_on_gap: Literal["fail", "probe", "clarify"] = Field(
-        description="The action to take when an epistemic gap is detected."
-    )
-
-
-class SelfCorrectionPolicy(CoreasonBaseModel):
-    """
-    Policy for self-correction and iterative refinement.
-    """
-
-    max_loops: int = Field(ge=0, le=50, description="The maximum number of self-correction loops allowed.")
-    rollback_on_failure: bool = Field(description="Whether to rollback to the previous state on failure.")
 
 
 class AgentNode(BaseNode):
