+++ src/coreason_manifest/workflow/policies.py
@@ -1,0 +1,46 @@
+# Copyright (c) 2025 CoReason, Inc.. All Rights Reserved
+#
+# This software is licensed under the Prosperity Public License 3.0.0.
+# The issuer of the Prosperity Public License for this software is CoReason, Inc..
+#
+# For a commercial version of this software, please contact us at gowtham.rao@coreason.ai.
+
+from typing import Literal
+
+from pydantic import Field
+
+from coreason_manifest.core.base import CoreasonBaseModel
+
+
+class System1Reflex(CoreasonBaseModel):
+    """
+    Policy for fast, intuitive system 1 thinking.
+    """
+
+    confidence_threshold: float = Field(
+        ge=0.0, le=1.0, description="The confidence threshold required to execute a reflex action."
+    )
+    allowed_read_only_tools: list[str] = Field(description="List of read-only tools allowed during a reflex action.")
+
+
+class EpistemicScanner(CoreasonBaseModel):
+    """
+    Policy for epistemic scanning and gap detection.
+    """
+
+    active: bool = Field(description="Whether the epistemic scanner is active.")
+    dissonance_threshold: float = Field(
+        ge=0.0, le=1.0, description="The threshold for cognitive dissonance before triggering an action."
+    )
+    action_on_gap: Literal["fail", "probe", "clarify"] = Field(
+        description="The action to take when an epistemic gap is detected."
+    )
+
+
+class SelfCorrectionPolicy(CoreasonBaseModel):
+    """
+    Policy for self-correction and iterative refinement.
+    """
+
+    max_loops: int = Field(ge=0, le=50, description="The maximum number of self-correction loops allowed.")
+    rollback_on_failure: bool = Field(description="Whether to rollback to the previous state on failure.")
